mock seachbar + tags
